savepoint

+++ vidly/routes/movies.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const movie = require('../models/movie')
+const movie = require('../models/movie');
+const genreModel = require('../models/genre');
 
 router.get('/', async (req, res) => {
 	const movies = await movie.Movie.find();
@@ -10,21 +11,29 @@ router.get('/', async (req, res) => {
 });
 
 router.get('/byId/:id', async (req, res) => {
-	const movie_id = req.params.id;
-	const result = await movie.Movie.findById(movie_id);
-	if (!result) return res.status(404).send('Movie not found.');
+	if ( mongoose.Types.ObjectId.isValid(req.params.id)) {
+		const movie_id = req.params.id;
+		const result = await movie.Movie.findById(movie_id);
+		if (!result) return res.status(404).send('Movie not found.');
 
-	res.send(result);
+		res.send(result);
+	}
+	else {
+		return res.status(400).send('Invalid Movie Id.');
+	}
 });
 
 router.post('/', async (req, res) => {
 	const { error } = movie.validateMovie(req.body);
 	if (error) return res.status(400).send(error);
+	const genre = await genreModel.Genre.findById(req.body.genreId);
+	if (!genre) return res.status(400).send('Genre not found.');
+
 	const m_obj = new movie.Movie({
 		title: req.body.title,
 		genre: {
-			name: req.body.genre.name,
-			isActive: req.body.genre.isActive
+			_id: genre._id,
+			name: genre.name,
 		},
 		numberInStock: req.body.numberInStock,
 		dailyRentalRate: req.body.dailyRentalRate
@@ -40,14 +49,24 @@ router.post('/', async (req, res) => {
 });
 
 router.put('/:id', async (req,res) => {
-	const movie_id = req.params.id;
-	const result = await movie.Movie.findOneAndUpdate(movie_id, req.body);
-	res.send(result);
+	if ( mongoose.Types.ObjectId.isValid(req.params.id)) {
+		const movie_id = req.params.id;
+		const result = await movie.Movie.findOneAndUpdate({ _id: movie_id }, req.body);
+		res.send(result);
+	}
+	else {
+		return res.status(400).send('Invalid Movie Id.');
+	}
 });
 
 router.delete('/:id', async (req, res) => {
-	const result = await movie.Movie.deleteOne({ _id: req.params.id });
-	res.send(result);
+	if ( mongoose.Types.ObjectId.isValid(req.params.id)) {
+		const result = await movie.Movie.deleteOne({ _id: req.params.id });
+		res.send(result);
+	}
+	else {
+		return res.status(400).send('Invalid Movie Id.');
+	}
 });
 
 module.exports = router;
